requirements.txt
streamlit
pandas
pyyaml
plotly
altair
networkx
pyvis
matplotlib
